rename env_params_new.py to env_params.py
