+++ echopype/calibrate/calibrate_azfp.py
@@ -3,7 +3,7 @@ import numpy as np
 from ..echodata import EchoData
 from .cal_params import get_cal_params_AZFP
 from .calibrate_ek import CalibrateBase
-from .env_params_new import get_env_params_AZFP
+from .env_params import get_env_params_AZFP
 from .range import compute_range_AZFP
