+++ echopype/calibrate/calibrate_ek.py
@@ -10,7 +10,7 @@ from ..utils.log import _init_logger
 from .cal_params import get_cal_params_EK, get_gain_BB, get_vend_filter_EK80
 from .calibrate_base import CalibrateBase
 from .ek80_complex import compress_pulse, get_tau_effective, get_transmit_signal
-from .env_params_new import get_env_params_EK60, get_env_params_EK80
+from .env_params import get_env_params_EK60, get_env_params_EK80
 from .range import compute_range_EK
 
 logger = _init_logger(__name__)
